core: Cache SSR results as long as possible
The results are dependent on:

- Request path
- CSRF token
- User

So this is mainly useful for when a given user navigates around a site
or reloads a page.

When loaders are implemented, this caching strategy will need to be
updated, but it should still work fairly well, assuming page load data
doesn't change very often.