Remove distinction between page & API/handler routes
Originally, API routes, prefixed with `$api`, were mainly intended to
distinguish between requests for a page versus requests for data for a
page, but that can be handled by checking the `Accept` header, which
seems more correct while also presenting nicer URLs to users (in those
cases where such URLs are exposed).

In this version, there are just routes and there's a single view for a
given route. For `GET` requests in particular, a route view will
determine whether to render a page or call the `get` handler based on
whether the `Accept` header is `application/json`.

This will also make it easier to inject loader data from the
associated `get` handler when rendering a page (or from the designated
loader).

This also:

- Uses "handler" everywhere instead of "api", "API", and "API handler"
  (including Python module names)
- Improves testing and test coverage
- Improves the `make_request()` utility function
- Renames the the `add-api` command to `add-handler` (and renames the
  related options on `add-page`)
- Replaces the `is_fetch` request attribute with a `prefers_json`
  attribute that is based on the request's `Accept` header (this seems
  more RESTful)