Use TOML files for settings instead of .env files
TOML files are conceptually similar to .env files, but they're *much*
easier to work with, especially when it comes to nested and non-str
values, and this commit removes a bunch of complex code related to
loaded and parsing env vars.

It's still possible to load specific settings from env vars by listing
them in the `[django.from_env]` and/or `[djangokit.from_env]` sections
of the TOML settings file. This is intended mainly for secret settings.

IDEA: Provide other `from_xyz` hooks to load settings from, e.g., a
      secrets manager.