core: Flesh out handlers & loaders
This contains a few related changes:

1. Handlers are now wrapped in a `Handler`, which contains the config
   for a handler.

2. Additional handlers can be added for a route using the associated
   `@handler` decorator.

3. `@handler` can also be used to configure caching for a handler.

   NOTE: This needs review and will probably need to be tweaked.

4. The handler that should be used as the loader for a route is now
   attached to the route's view.

   TODO: The loader is being called for SSR now and the data is being
         passed into the SSR bundle, but the data isn't being used yet.

This also simplifies the URLconf for catchall routes by using
`<path:path>`. The `path` converter consumes whatever is left of the
URL path just like the `.*` regex, and this allows all URLconfs to be
configured using `path` rather than having to check for a regex pattern
and use `re_path` for the catchall special case.