website: Deploy static files to shared remote directory
Rather than creating a new static directory for each deployment, use
a shared directory so that static files for previous versions aren't
immediately removed (or, more precisely, made inaccessible).

This avoids missing static files when a user has a previous version of
a page in their local cache, which can cause completely unstyled pages
to be displayed, JS files to not be loaded, etc.