core: Move DjangoKitSettings dataclass to its own module
This is so it can be used separately without having to do all the
settings loading stuff in `djangokit.core.settings`.

This supports projects where the user wants to use a regular Django
settings module. DjangoKitSettings needs to be used in these cases so
the DK settings are properly initialized.